+++ src/utils/plist.h
@@ -29,13 +29,13 @@ uint8_t plist_object_get_type(const plist_object_t *object);
 int plist_object_primitive_get_value(const plist_object_t *object, uint8_t *value);
 int plist_object_integer_get_value(const plist_object_t *object, int64_t *value);
 int plist_object_real_get_value(const plist_object_t *object, double *value);
-int plist_object_data_get_value(const plist_object_t *object, const uint8_t **value, uint32_t *valuelen);
+int plist_object_data_get_value(const plist_object_t *object, const uint8_t **value, uint64_t *valuelen);
 int plist_object_string_get_value(const plist_object_t *object, const char **value);
 const plist_object_t *plist_object_array_get_value(const plist_object_t *object, uint32_t idx);
 const plist_object_t *plist_object_dict_get_value(const plist_object_t *object, const char *key);
 
 plist_object_t *plist_object_from_bplist(const uint8_t *data, uint32_t datalen);
-int plist_object_to_bplist(plist_object_t *object, uint8_t **data, uint32_t *datalen);
+int plist_object_to_bplist(plist_object_t *object, uint8_t **data, uint64_t *datalen);
 
 void plist_object_destroy(plist_object_t *object);
 
@@ -102,7 +102,7 @@ public:
     {
         std::vector<uint8_t> buf;
         uint8_t *data = 0;
-        uint32_t length = 0;
+        uint64_t length = 0;
         plist_object_to_bplist(obj_, &data, &length);
         buf.assign(data, data + length);
         free(data);
